FirebaseService now implemented
