+++ app/src/main/java/com/followapp/mytasks/loginModule/model/LoginRepository.kt
@@ -9,18 +9,17 @@ import androidx.credentials.CustomCredential
 import androidx.credentials.GetCredentialRequest
 import androidx.credentials.GetCredentialResponse
 import com.followapp.mytasks.R
+import com.followapp.mytasks.common.dataAccess.services.FirebaseService
+
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import com.google.android.libraries.identity.googleid.GoogleIdTokenParsingException
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.GoogleAuthProvider
 
-class LoginRepository {
+class LoginRepository(private val firebaseService: FirebaseService) {
 
-    private val auth: FirebaseAuth = FirebaseAuth.getInstance()
     private lateinit var credentialManager: CredentialManager
 
     fun checkPlayServices(context: Context): Boolean {
@@ -46,61 +45,46 @@ class LoginRepository {
         }
     }
 
-suspend fun getGoogleIdToken(context: Context): GetCredentialResponse? {
-    return try {
-        credentialManager = CredentialManager.create(context)
-        val googleIdOption = GetGoogleIdOption.Builder()
-            .setFilterByAuthorizedAccounts(false)
-            .setServerClientId(context.getString(R.string.web_client_id))
-            .build()
+    suspend fun getGoogleIdToken(context: Context): GetCredentialResponse? {
+        return try {
+            credentialManager = CredentialManager.create(context)
+            val googleIdOption = GetGoogleIdOption.Builder()
+                .setFilterByAuthorizedAccounts(false)
+                .setServerClientId(context.getString(R.string.web_client_id))
+                .build()
 
-        val request = GetCredentialRequest.Builder()
-            .addCredentialOption(googleIdOption)
-            .build()
+            val request = GetCredentialRequest.Builder()
+                .addCredentialOption(googleIdOption)
+                .build()
 
-        credentialManager.getCredential(context, request)
-    } catch (e: Exception) {
-        Log.e("LoginRepository", "Error getting Google ID token", e)
-        null
+            credentialManager.getCredential(context, request)
+        } catch (e: Exception) {
+            Log.e("LoginRepository", "Error getting Google ID token", e)
+            null
+        }
     }
-}
 
-    fun handleSignIn(result: GetCredentialResponse, callback: (FirebaseUser?, String?) -> Unit) {
+    private fun handleSignIn(result: GetCredentialResponse, callback: (FirebaseUser?, String?) -> Unit) {
         when (val credential = result.credential) {
             is GoogleIdTokenCredential -> {
-                try {
-                    firebaseAuthWithGoogle(credential.idToken, callback)
-                } catch (e: GoogleIdTokenParsingException) {
-                    callback(null, e.message)
-                }
+                firebaseService.handleSignIn(credential, callback)
             }
 
             is CustomCredential -> {
                 if (credential.type == GoogleIdTokenCredential.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
                     try {
                         val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
-                        firebaseAuthWithGoogle(googleIdTokenCredential.idToken, callback)
+                        firebaseService.handleSignIn(googleIdTokenCredential, callback)
                     } catch (e: GoogleIdTokenParsingException) {
-                        Log.e("LoginFragment", "Received an invalid google id token response", e)
+                        Log.e("LoginRepository", "Received an invalid google id token response", e)
                         callback(null, e.message)
                     }
                 }
             }
 
             else -> {
-                Log.e("LoginFragment", "Unexpected type of credential: ${credential::class.java}")
+                Log.e("LoginRepository", "Unexpected type of credential: ${credential::class.java}")
                 callback(null, "Unexpected type of credential")
-            }
-        }
-    }
-
-    private fun firebaseAuthWithGoogle(idToken: String, callback: (FirebaseUser?, String?) -> Unit) {
-        val credential = GoogleAuthProvider.getCredential(idToken, null)
-        auth.signInWithCredential(credential).addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                callback(auth.currentUser, null)
-            } else {
-                callback(null, task.exception?.message)
             }
         }
     }
