+++ app/src/main/java/com/followapp/mytasks/loginModule/view/LoginFragment.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.followapp.mytasks.R
+import com.followapp.mytasks.common.dataAccess.services.FirebaseService
 import com.followapp.mytasks.databinding.FragmentLoginBinding
 import com.followapp.mytasks.homeModule.view.HomeFragment
 import com.followapp.mytasks.loginModule.model.LoginRepository
@@ -36,7 +37,7 @@ class LoginFragment : Fragment() {
     }
 
     private fun setupViewModel() {
-        loginViewModel = ViewModelProvider(this, LoginViewModelFactory(LoginRepository()))[LoginViewModel::class.java]
+        loginViewModel = ViewModelProvider(this, LoginViewModelFactory(LoginRepository(FirebaseService())))[LoginViewModel::class.java]
     }
 
     private fun setupObservers() {
